moved ProgressBarDialog to a new file, added ctrl_type field to ControlTreeNode
